move fetch-news into separate action
