+++ native/cli/src/action.rs
@@ -1,3 +1,5 @@
 mod post_news;
+mod fetch_news;
 
 pub use post_news::PostNews;
+pub use fetch_news::FetchNews;
